Fix build steps in GitHub Actions workflow
The `deepsweep` workflow was not running the build scripts, causing a race condition where a deployment could happen with stale or empty HTML files.

- Added steps to `.github/workflows/psai-deepsweep.yml` to run the build scripts for the feeds and sources pages.
- Updated the commit step in the workflow to include the generated public files.
