+++ .github/workflows/psai-deepsweep.yml
@@ -48,11 +48,28 @@ jobs:
         run: |
           python scripts/merge_candidates.py             --candidates data/candidates.json             --tracker data/PSAI_FullField_Tracker.csv             --out data/PSAI_FullField_Tracker.csv
 
+      - name: Build feeds
+        env:
+          SITE_URL: ${{ vars.SITE_URL }}
+        run: |
+          SITE="${SITE_URL}"
+          if [ -z "$SITE" ]; then
+            SITE="https://${GITHUB_REPOSITORY_OWNER}.github.io/${GITHUB_REPOSITORY#*/}"
+          fi
+          python scripts/build_feed_from_log.py --site "$SITE"
+
+      - name: Build sources pages (cards + table)
+        env:
+          PSAI_INCLUDE_REGEX: "(agent|agentic|orchestr|mcp|ide|editor|review|code\\s*assistant)"
+          PSAI_EXCLUDE_REGEX: "^(product\\s*hunt|reddit|hacker\\s*news|hn\\s*—|hn\\s*show|latentspace|ben['’]s\\s*bites|npm|pypi|docker\\s*hub)\\b"
+          PSAI_MIN_STARS: "50"
+        run: python scripts/build_sources_pages.py
+
       - name: Commit discovery results
         run: |
           git config user.name "github-actions[bot]"
           git config user.email "github-actions[bot]@users.noreply.github.com"
-          git add data/candidates.json data/PSAI_FullField_Tracker.csv || true
+          git add data/candidates.json data/PSAI_FullField_Tracker.csv public/*.html public/*.json public/*.xml || true
           git commit -m "PSAI DeepSweep: discovery & tracker append $(date -u +'%Y-%m-%dT%H:%M:%SZ')" || echo "No changes"
           git push || true
