+++ public/index.html
@@ -49,7 +49,7 @@ th{text-align:left;background:#f8fafc;font-weight:600}
 <main>
 
   <h1>PSAI Dashboard</h1>
-  <div class="sub">Last updated: 2025-08-14 01:46:25 UTC</div>
+  <div class="sub">Last updated: 2025-08-14 02:04:53 UTC</div>
   <div class="grid" style="grid-template-columns:1fr 1fr;max-width:1200px;">
     <section class="card">
       <h2>Today's Updates</h2>
